+++ keystatic/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @keystatic/test-next-app
 
+## 0.0.38
+
+### Patch Changes
+
+- Updated dependencies [59a6bcf]
+  - @keystatic/core@0.0.68
+
+## 0.0.37
+
+### Patch Changes
+
+- Updated dependencies [5f9a6c8]
+  - @keystatic/core@0.0.67
+
 ## 0.0.36
 
 ### Patch Changes