+++ packages/keystatic/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @keystatic/core
 
+## 0.0.68
+
+### Patch Changes
+
+- 59a6bcf: Use port that Keystatic is running on when doing setup for callback
+  URLs
+
+## 0.0.67
+
+### Patch Changes
+
+- 5f9a6c8: Remove usage of `administration` permissions in forking flow
+
 ## 0.0.66
 
 ### Patch Changes